[MODORDSTOR-381] Merge with master and refactor code
