+++ src/main/java/org/folio/rest/impl/RoutingListsAPI.java
@@ -8,7 +8,7 @@ import org.folio.rest.annotations.Validate;
 import org.folio.rest.core.models.RequestContext;
 import org.folio.rest.jaxrs.model.RoutingList;
 import org.folio.rest.jaxrs.resource.OrdersRoutingLists;
-import org.folio.service.routinglists.RoutingListsService;
+import org.folio.service.routinglists.RoutingListService;
 import org.folio.spring.SpringContextUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -20,7 +20,7 @@ import static io.vertx.core.Future.succeededFuture;
 public class RoutingListsAPI extends BaseApi implements OrdersRoutingLists {
 
   @Autowired
-  private RoutingListService routingListsService;
+  private RoutingListService routingListService;
 
   public RoutingListsAPI() {
     SpringContextUtil.autowireDependencies(this, Vertx.currentContext());
@@ -29,7 +29,7 @@ public class RoutingListsAPI extends BaseApi implements OrdersRoutingLists {
   @Override
   @Validate
   public void getOrdersRoutingLists(String query, String totalRecords, int offset, int limit, Map<String, String> okapiHeaders, Handler<AsyncResult<Response>> asyncResultHandler, Context vertxContext) {
-    routingListsStorageService.getRoutingLists(limit, offset, query, new RequestContext(vertxContext, okapiHeaders))
+    routingListService.getRoutingLists(limit, offset, query, new RequestContext(vertxContext, okapiHeaders))
       .onSuccess(lists -> asyncResultHandler.handle(succeededFuture(buildOkResponse(lists))))
       .onFailure(fail -> handleErrorResponse(asyncResultHandler, fail));
   }
@@ -37,7 +37,7 @@ public class RoutingListsAPI extends BaseApi implements OrdersRoutingLists {
   @Override
   @Validate
   public void postOrdersRoutingLists(RoutingList entity, Map<String, String> okapiHeaders, Handler<AsyncResult<Response>> asyncResultHandler, Context vertxContext) {
-    routingListsStorageService.createRoutingList(entity, new RequestContext(vertxContext, okapiHeaders))
+    routingListService.createRoutingList(entity, new RequestContext(vertxContext, okapiHeaders))
       .onSuccess(list -> asyncResultHandler.handle(succeededFuture(buildOkResponse(list))))
       .onFailure(fail -> handleErrorResponse(asyncResultHandler, fail));
   }
@@ -45,7 +45,7 @@ public class RoutingListsAPI extends BaseApi implements OrdersRoutingLists {
   @Override
   @Validate
   public void getOrdersRoutingListsById(String id, Map<String, String> okapiHeaders, Handler<AsyncResult<Response>> asyncResultHandler, Context vertxContext) {
-    routingListsStorageService.getRoutingList(id, new RequestContext(vertxContext, okapiHeaders))
+    routingListService.getRoutingList(id, new RequestContext(vertxContext, okapiHeaders))
       .onSuccess(list -> asyncResultHandler.handle(succeededFuture(buildOkResponse(list))))
       .onFailure(fail -> handleErrorResponse(asyncResultHandler, fail));
   }
@@ -53,7 +53,7 @@ public class RoutingListsAPI extends BaseApi implements OrdersRoutingLists {
   @Override
   @Validate
   public void deleteOrdersRoutingListsById(String id, Map<String, String> okapiHeaders, Handler<AsyncResult<Response>> asyncResultHandler, Context vertxContext) {
-    routingListsStorageService.deleteRoutingList(id, new RequestContext(vertxContext, okapiHeaders))
+    routingListService.deleteRoutingList(id, new RequestContext(vertxContext, okapiHeaders))
       .onSuccess(list -> asyncResultHandler.handle(succeededFuture(buildNoContentResponse())))
       .onFailure(fail -> handleErrorResponse(asyncResultHandler, fail));
 
@@ -62,7 +62,7 @@ public class RoutingListsAPI extends BaseApi implements OrdersRoutingLists {
   @Override
   @Validate
   public void putOrdersRoutingListsById(String id, RoutingList entity, Map<String, String> okapiHeaders, Handler<AsyncResult<Response>> asyncResultHandler, Context vertxContext) {
-    routingListsStorageService.updateRoutingList(entity, new RequestContext(vertxContext, okapiHeaders))
+    routingListService.updateRoutingList(entity, new RequestContext(vertxContext, okapiHeaders))
       .onSuccess(list -> asyncResultHandler.handle(succeededFuture(buildOkResponse(list))))
       .onFailure(fail -> handleErrorResponse(asyncResultHandler, fail));
   }
