+++ src/test/java/org/folio/ApiTestSuite.java
@@ -99,7 +99,7 @@ import org.folio.service.pieces.flows.update.PieceUpdateFlowInventoryManagerTest
 import org.folio.service.pieces.flows.update.PieceUpdateFlowManagerTest;
 import org.folio.service.pieces.flows.update.PieceUpdateFlowPoLineServiceTest;
 import org.folio.service.pieces.validators.PieceValidatorUtilTest;
-import org.folio.service.RoutingListServiceTest;
+import org.folio.service.routinglists.RoutingListServiceTest;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Nested;
