+++ src/test/java/org/folio/rest/impl/RoutingListsApiTest.java
@@ -13,6 +13,7 @@ import org.folio.rest.jaxrs.model.Error;
 import org.folio.rest.jaxrs.model.Errors;
 import org.folio.rest.jaxrs.model.PoLine;
 import org.folio.service.orders.PurchaseOrderLineService;
+import org.folio.service.routinglists.RoutingListService;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
@@ -48,6 +49,7 @@ import static org.folio.TestConstants.EMPTY_CONFIG_X_OKAPI_TENANT;
 import static org.folio.TestConstants.EXIST_CONFIG_X_OKAPI_TENANT_LIMIT_10;
 import static org.folio.TestConstants.ID_BAD_FORMAT;
 import static org.folio.TestConstants.ID_DOES_NOT_EXIST;
+import static org.folio.TestConstants.ROUTING_LIST_ID;
 import static org.folio.TestConstants.X_OKAPI_TOKEN;
 import static org.folio.TestConstants.X_OKAPI_USER_ID;
 import static org.folio.TestConstants.X_OKAPI_USER_ID_WITH_ACQ_UNITS;
